feat: add group layouts for main and special pages
