+++ src/routes/(main)/+page.svelte
