+++ src/routes/(special)/maintenance/+layout.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+	import '$app.css';
+</script>
+
+<div class="w-full">
+	<slot />
+</div>
