+++ src/routes/(special)/maintenance/+page.svelte
@@ -1,0 +1,5 @@
+<section class="flex items-center justify-center h-screen w-full bg-red-700">
+	<div class="block">
+		<span class="font-integralCF text-8xl tracking-wide text-white">New website cooming soon</span>
+	</div>
+</section>
