+++ svelte.config.js
@@ -2,10 +2,18 @@ import adapter from '@sveltejs/adapter-static';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
+
 export default {
   preprocess: vitePreprocess(),
 
 	kit: {
+    alias: {
+      $types: 'src/types.ts',
+      '$app.css': 'src/app.css',
+      $utils: 'src/utils',
+      $components: 'src/components',
+      $icons: 'src/icons'
+    },
 		adapter: adapter({
 			// default options are shown. On some platforms
 			// these options are set automatically — see below
